Moved appcatalog's Initialize.java to proper folder
